requirements.txt
pymongo>=4.6.0
pandas>=2.1.0
python-dotenv>=1.0.0
dnspython>=2.0.0